requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.30.0
alembic==1.12.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Task queue
celery==5.3.4
redis==5.0.1

# AI / ML
sentence-transformers==3.0.1
qdrant-client==1.7.3
groq==0.4.2

# Document processing
pymupdf==1.25.3

# Config & utilities
pydantic==2.10.0
pydantic-settings==2.6.1
python-dotenv==1.0.0
httpx==0.25.2
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
